finished

+++ app/Http/Controllers/ListingController.php
@@ -43,6 +43,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created succesfully');
@@ -50,11 +52,19 @@ class ListingController extends Controller
 
     // Show edit form
     public function edit(Listing $listing) {
+        // make sure current user is authorized
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         return view('listings.edit', ['listing' => $listing]);
     }
 
     // Update listing data
     public function update(Request $request, Listing $listing) {
+        // make sure current user is authorized
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -75,7 +85,16 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        // make sure current user is authorized
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // manage listing
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
